ptfs: switch to new implementation from util.rs
1. switch to new implementation from util.rs.
2. move helpers into util.rs
3. split out config.rs for configuration information.
4. move os compatibility related data structs into os_compat.rs
